LayoutTests/http/tests/security/xss-DENIED-non-shadowable-propterty-with-iframe-proto-expected.txt


Tests that making other frame window a prototype doesn't expose that window properties
PASS: location === originalLocation should be 'true' and is.
PASS: this.location === originalLocation should be 'true' and is.
